requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
pyyaml>=6.0
rich>=13.0.0
python-dotenv>=1.0.0
fastapi>=0.110.0
uvicorn>=0.29.0
